requirements.txt
pillow
gradio
ultralytics
torch
numpy